refactor(dashboard): relocate dock-zone model to src/ as Neo.dashboard.DockZoneModel (#13120)
Operator-directed (the PR #13116 review): dock zones are a reusable Neo layout topic, available to other apps, so they belong in src/. Moves AgentOS.dock.DockZoneModel (apps/agentos/dock/) -> Neo.dashboard.DockZoneModel (src/dashboard/), co-located with the already-merged Neo.dashboard.DockLayoutAdapter + the dashboard subsystem. Pure relocation, no behavior change: the 30 unit tests pass unchanged at the new location, and no code imports the model (it is contract-consumed via the .md doc), so DockPreview and the adapter are untouched.

Reconciles the HarnessDockZoneModel.md Ownership Boundary inconsistency (the conditional 'harness app layer' model placement + 'second-consumer-required' gate vs the 'dashboard layer' statement) to the dashboard layer, per the operator's decision.
